#130: Change the output page to be similar to the one on the desktop wallet
