+++ src/app/components/pages/wallets/address-detail/wallet-detail.component.html
@@ -3,10 +3,12 @@
     <div class="id-column">{{ i + 1 }}</div>
     <div class="address-column">
       <img src="../../../../../assets/img/qr-code-black.png" (click)="showQr(address)" >
-      <p class="click-to-copy" [ngClass]="{ copying: address.isCopying }" (click)="copyAddress(address)">
+      <p class="click-to-copy" [ngClass]="{ copying: address.isCopying }" (click)="onCopyAddress(address)">
         {{ address.address }} <span data-label="Copied" class="copy-label">Copy</span>
       </p>
     </div>
+    <button class="unspent-outputs-column" mat-button
+            routerLink="/settings/outputs" [queryParams]="{ addr: address.address }">Unspent Outputs</button>
     <div class="hours-column">{{ address.hours ? address.hours : 0 }}</div>
     <div class="coins-column">{{ (address.balance ? address.balance : 0) | number:'1.0-6' }}</div>
   </div>
