+++ src/app/components/pages/wallets/address-detail/wallet-detail.component.scss
